Loops refactored.
