+++ gamestate/loop/pause_menu_loop.py
@@ -4,44 +4,16 @@ import os
 from config import menu_data
 from events import listen
 
-from gamestate.loop.loop import Loop
+from gamestate.loop.menu_loop import MenuLoop
 from gamestate.menu.pause_menu import PauseMenu
 
 
-class PauseMenuLoop(Loop):
+class PauseMenuLoop(MenuLoop):
     def __init__(self):
         super().__init__()
-        self.load_assets()
-        self.time_since_loop_started = pygame.time.get_ticks()
-        self.can_interact = False
-
-    def draw(self):
-        self.menu.draw_background()
-        self.menu.draw_text_sprites()
-        self.menu.draw_button_sprites()
 
     def load_assets(self):
         self.menu = PauseMenu(self.screen, menu_data.PAUSE_MENU)
-
-    def update(self):
-        pygame.display.update()
-
-    def run(self):
-        self.time_since_loop_started = pygame.time.get_ticks()
-
-        self.running = True
-        while self.running:
-            self.cooldown()
-            self.event_loop()
-            self.update()
-            self.draw()
-
-    def cooldown(self):
-        current_time = pygame.time.get_ticks()
-        delta_time = current_time - self.time_since_loop_started
-
-        if delta_time >= 500:
-            self.can_interact = True
 
     def event_loop(self):
         listen.event_loop()
@@ -50,5 +22,4 @@ class PauseMenuLoop(Loop):
             self.menu.check_events()
             pressed_keys = pygame.key.get_pressed()
             if pressed_keys[pygame.K_ESCAPE]:
-                if self.can_interact:
-                    self.running = False
+                self.running = False
